api/requirements.txt
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Cors==3.0.8
Flask-Gravatar==0.5.0
Flask-HTMLmin==1.5.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-MySQL==1.5.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
flask-restplus==0.13.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
PyYAML==5.3
sqlparse==0.3.0
sshtunnel==0.1.5
typed-ast==1.4.1
watchdog==0.9.0
webencodings==0.5.1
Werkzeug==0.16.0
WTForms==2.2.1
pymongo==3.11.3
greenlet==1.1.2
importlib-metadata==4.8.1
typing-extensions==3.10.0.2
zipp==3.6.0
pycryptodome==3.11.0
napalm==3.3.1
pysnmp==4.4.10